+++ src/com/luoyk/toolbox/panel/Redis.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="panel.WebSocekt">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="com.luoyk.toolbox.panel.Redis">
   <grid id="27dc6" binding="panel" layout-manager="GridLayoutManager" row-count="2" column-count="2" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
     <margin top="0" left="0" bottom="0" right="0"/>
     <constraints>
@@ -8,20 +8,20 @@
     <properties/>
     <border type="none"/>
     <children>
-      <component id="7053c" class="javax.swing.JLabel">
+      <component id="d0695" class="javax.swing.JLabel">
         <constraints>
           <grid row="0" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
         </constraints>
         <properties>
-          <text value="websocket"/>
+          <text value="redis"/>
         </properties>
       </component>
-      <hspacer id="5e22c">
+      <hspacer id="d4b4b">
         <constraints>
           <grid row="0" column="1" row-span="1" col-span="1" vsize-policy="1" hsize-policy="6" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
         </constraints>
       </hspacer>
-      <vspacer id="751f3">
+      <vspacer id="70cf6">
         <constraints>
           <grid row="1" column="0" row-span="1" col-span="1" vsize-policy="6" hsize-policy="1" anchor="0" fill="2" indent="0" use-parent-layout="false"/>
         </constraints>
